highlights internal parts of scala.reflect.runtime
By moving them to scala.reflect.runtime.internal.

Noone should be using those classes anyway, because everything except
JavaUniverse is private[XXX] and JavaUniverse itself has the "do not
instantiate" doc comment on it.

On the other hand, our runtime reflection APIs are exposed via
api.JavaUniverse, not via runtime.JavaUniverse, so nothing publicly
accessible refers to the runtime cake.

Finally, reifier is also not using runtime cake directly, preferring
api.JavaUniverse. It didn't have to, but I feel very relieved that I did
not tap into internal implementation details back then. Here's the code:
https://github.com/scala/scala/blob/v2.10.0/src/compiler/scala/reflect/reify/package.scala#L37
