+++ src/compiler/scala/tools/nsc/scratchpad/Mixer.scala
@@ -4,7 +4,7 @@ import java.io.{FileInputStream, InputStreamReader, IOException}
 
 import scala.runtime.ScalaRunTime.stringOf
 import java.lang.reflect.InvocationTargetException
-import scala.reflect.runtime.ReflectionUtils._
+import scala.reflect.runtime.internal.ReflectionUtils._
 import scala.collection.mutable.ArrayBuffer
 
 @deprecated("SI-6458: Instrumentation logic will be moved out of the compiler.","2.10.0")
