+++ src/reflect/scala/reflect/runtime/internal/ReflectSetup.scala
@@ -1,11 +1,10 @@
-package scala.reflect
-package runtime
+package scala.reflect.runtime.internal
 
-import internal.{SomePhase, NoPhase, Phase, TreeGen}
+import scala.reflect.internal.{SomePhase, NoPhase, Phase, TreeGen}
 
 /** A helper trait to initialize things that need to be set before JavaMirrors and other
  *  reflect specific traits are initialized */
-private[runtime] trait ReflectSetup extends internal.SymbolTable {
+private[runtime] trait ReflectSetup extends scala.reflect.internal.SymbolTable {
   override val phaseWithId: Array[Phase] = Array(NoPhase, SomePhase)
   override val currentRunId = 1 // fake a run id so that it is different from NoRunId
   phase = SomePhase // set to a phase different from NoPhase
